BREAKING CHANGE: upgrade to gradle 8.2, and remove kotlin mpp/js support
