--- gradle.properties
@@ -1,4 +1,0 @@
-kotlin.code.style=official
-kotlin.mpp.enableGranularSourceSetsMetadata=true
-kotlin.native.enableDependencyPropagation=false
-kotlin.js.generate.executable.default=false
