--- jitpack.yml
@@ -1,4 +1,0 @@
-jdk:
-  - openjdk17
-install:
-  - ./gradlew allTests publishToMavenLocal
