+++ settings.gradle.kts
@@ -1,1 +1,5 @@
-rootProject.name = "aff-compose"
+plugins {
+    id("org.gradle.toolchains.foojay-resolver-convention") version "0.7.0"
+}
+
+rootProject.name = "AffCompose"
