--- src/jvmTest/kotlin/com/tairitsu/compose/arcaea/misc.kt
@@ -1,14 +1,0 @@
-package com.tairitsu.compose.arcaea
-
-fun assertDifficultyEquals(a: Difficulty?, b: Difficulty?) {
-    org.junit.jupiter.api.Assertions.assertNotNull(a)
-    org.junit.jupiter.api.Assertions.assertNotNull(b)
-
-    a!!
-    b!!
-
-    org.junit.jupiter.api.Assertions.assertEquals(a.chartDesigner, b.chartDesigner)
-    org.junit.jupiter.api.Assertions.assertEquals(a.jacketDesigner, b.jacketDesigner)
-    org.junit.jupiter.api.Assertions.assertEquals(a.rating, b.rating)
-    org.junit.jupiter.api.Assertions.assertEquals(a.ratingClass, b.ratingClass)
-}
