+++ src/main/kotlin/com/tairitsu/compose/arcaea/JvmCompose.kt
@@ -1,0 +1,24 @@
+package com.tairitsu.compose.arcaea
+
+import io.ktor.utils.io.streams.*
+import java.io.File
+import java.io.FileOutputStream
+import java.nio.file.Files
+import java.nio.file.Path
+
+fun MapSet.writeToFolder(outputPath: Path) {
+    writeToFolder(outputPath.toFile())
+}
+
+fun MapSet.writeToFolder(outputPath: File) {
+    Files.createDirectories(outputPath.toPath())
+    writeToOutput { fileName ->
+        FileOutputStream(File(outputPath, fileName)).asOutput()
+    }
+}
+
+fun Difficulty.printToConsole() {
+    this.chart.serialize().let {
+        println(it)
+    }
+}
